updated test for sad paths #cryptidbyid
