+++ spec/graphql/queries/cryptids/cryptid_by_id_spec.rb
@@ -14,6 +14,21 @@ module Queries
         expect(result['data']['cryptidById']['range']).to be_a String
         expect(result['data']['cryptidById']['image']).to be_a String
       end
+
+      it 'returns an error if no cryptid by id' do
+        create(:cryptid, id: '2')
+
+        result = OkCryptidBeSchema.execute(cryptid_by_id_query).as_json
+
+        expect(result).to have_key("errors")
+      end
+
+      it 'returns an error if cryptid idis left blank' do
+        create(:cryptid, id: '2')
+
+        result = OkCryptidBeSchema.execute(cryptid_by_id_query_id_blank).as_json
+        expect(result).to have_key("errors")
+      end
     end
   end
 end
@@ -22,6 +37,20 @@ def cryptid_by_id_query
   <<~GQL
     {
       cryptidById(id: "1") {
+        name
+        description
+        dangerLevel
+        range
+        image
+        }
+    }
+  GQL
+end
+
+def cryptid_by_id_query_id_blank
+  <<~GQL
+    {
+      cryptidById(id: "") {
         name
         description
         dangerLevel
